Rename `esrally.client.client` to `esrally.client.factory`.
