+++ tests/client/factory_test.py
@@ -64,7 +64,7 @@ class TestEsClientFactory:
         # make a copy so we can verify later that the factory did not modify it
         original_client_options = deepcopy(client_options)
 
-        logger = logging.getLogger("esrally.client.client")
+        logger = logging.getLogger("esrally.client.factory")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.EsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls(
@@ -105,7 +105,7 @@ class TestEsClientFactory:
         # make a copy so we can verify later that the factory did not modify it
         original_client_options = deepcopy(client_options)
 
-        logger = logging.getLogger("esrally.client.client")
+        logger = logging.getLogger("esrally.client.factory")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.EsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls(
@@ -147,7 +147,7 @@ class TestEsClientFactory:
         # make a copy so we can verify later that the factory did not modify it
         original_client_options = deepcopy(client_options)
 
-        logger = logging.getLogger("esrally.client.client")
+        logger = logging.getLogger("esrally.client.factory")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.EsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls(
@@ -218,7 +218,7 @@ class TestEsClientFactory:
         # make a copy so we can verify later that the factory did not modify it
         original_client_options = dict(client_options)
 
-        logger = logging.getLogger("esrally.client.client")
+        logger = logging.getLogger("esrally.client.factory")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.EsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls(
@@ -259,7 +259,7 @@ class TestEsClientFactory:
         # make a copy so we can verify later that the factory did not modify it
         original_client_options = deepcopy(client_options)
 
-        logger = logging.getLogger("esrally.client.client")
+        logger = logging.getLogger("esrally.client.factory")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.EsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls(
